создание User, Ads, Image

+++ example-for-graduate-work-main/src/main/java/ru/skypro/homework/service/ImageService.java
@@ -6,5 +6,5 @@ import ru.skypro.homework.entity.Image;
 public interface ImageService {
     byte[] updateImage(Integer adsId, MultipartFile file);
 
-    Image getImageById(Long imageId);
+    Image getImageById(Integer imageId);
 }
